requirements.txt
# inference
torch==2.7.1
torchvision==0.22.1
transformers==4.49.0
diffusers[torch]==0.33.1
torchmetrics
einops
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
ipdb

# preprocess
fire
h5py
kornia
opencv-python
ipython==8.20.0
timm
scipy
json_tricks
# open3d

# download dataset
huggingface_hub
datasets
pyarrow
pandas
pycolmap
